+++ frontend/src/App.js
@@ -1,9 +1,10 @@
 import React, { Component } from 'react';
 import './assets/bootstrap.min.css';
 import { connect } from 'react-redux';
-import HomePage from './root/components';
-import  CategoryView, {CategoriesMenu}  from "./categories/components";
-import { PostView } from "./posts/components";
+import HomePage from './root/components/HomePage';
+import  CategoryView  from "./categories/components/CategoryView";
+import  CategoriesMenu  from "./categories/components/CategoriesMenu";
+import  PostView  from "./posts/components/PostView";
 import { Route, withRouter } from 'react-router-dom';
 import { getPosts } from "./posts/actions";
 import { getCategories } from './categories/actions';
@@ -21,7 +22,7 @@ class App extends Component {
       <div className="App">
         <CategoriesMenu />
         <div className="container">
-          <Route exact path="/" component={HomePage}/>
+          <Route exact path="/" component={ HomePage }/>
 
           <Route exact path="/category/:categoryPath" component={ CategoryView }/>
